test: load font at toplevel main.lua and print test timer
The test framework depends on the font so it should be loaded by the framework rather than being dependent on the graphics test to load it. As a result the current graphics print test doesn't make a lot of sense as it's actually a keyboard/mouse test and not a graphics print test at all. It can be improved upon separately.

Showing the timeout - it seemed like a nice courtesy and could be helpful for certain troubleshooting purposes.