getWorkingDirectory: implemented to return "./"
It's enough to allow most things that might want to use it to work. The function should only be used for local development or debugging purposes anyway, as the concept of CWD is not portable and is not friendly to end users running program via high-level shells (eg. Windows Explorer or Finder or such).