make: use compiler switches to control warnings/errors behavior
 - Improves printf verification for lutro assertions when using clang compiler
 - Consolidates warning behavior into makefile and vcxproj file
 - Reduced impact: disables only one specific warning which is deemed "generally harmless" (and which is specific to clang): use of the gnu varadic macro extension
 - All other warnings are enabled.

This removes dependency on the compiler switch `-include`. It was deemed low-value to the needs of the project to introduce a dependency on the switch, especially given the generally low usage rate of Visual Studio and Msbuild compiler or IDE among libretro users. Moreover, the Visual Studio's printf warnings have been moved into the project file using the VS2022 /MX: feature, however the promotion to error itself currently does nothing: it conflicts with the printf redirection macro which is used to surface console logs to the VS output window. This behavior may change in the future.

Note that printf formatting verification is not supported on the gnuc compiler. It should not be impacted by these changes.